requirements.txt
openpyxl>=3.1.0
pyyaml>=6.0
genson>=1.3.0
